Refactor FileSystem tests into separate files per command
Split monolithic FileSystemTests.cs into focused test files:
- GetContentTests.cs: Tests for reading file content
- GetChildItemTests.cs: Tests for listing directories
- RemoveItemTests.cs: Tests for removing files/directories
- GetLocationTests.cs: Tests for getting/setting current directory
- AliasTests.cs: Tests for Bash-style aliases (Cat, Ls, Pwd, Cd, Rm)
- RunAll.cs: Main runner to execute all test suites

This improves maintainability by keeping each test file focused on a single command's functionality.
